==> 07_AUROC_plots.Rmd <==
---
title: "Plot AUROC markers"
author: "Jordi Camps"
date: "2018 M09 24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(readxl)
library(dplyr)
library(ggplot2)
```

## Import excel file
```{r}
df <- read_excel("tables/k7_de_genes.xlsx")
```

## Name clusters
```{r}
df$sc3_7_markers_clusts <- factor(df$sc3_7_markers_clusts)
df$sc3_7_markers_clusts <-  plyr::revalue(df$sc3_7_markers_clusts, c("1" = "ISC 3", "2" = "ISC 1", "3" = "ISC 2", "7" = "Fibroblast", "6" = "SMC", "4" = "Satellite\ncell", "5" = "Schwann\ncell"))
df$sc3_7_markers_clusts <- factor(df$sc3_7_markers_clusts, levels = c("ISC 1", "ISC 2", "ISC 3", "Fibroblast", "SMC", "Satellite\ncell", "Schwann\ncell"))
```

## Select top 10 markers per cluster and order by highest value of AUROC
```{r fig.height=4, fig.width=5}
temp <- df %>%
  group_by(sc3_7_markers_clusts) %>%
  #filter(sc3_7_markers_clusts == 2) %>%
  arrange(desc(sc3_7_markers_auroc)) %>%
  top_n(10, wt = sc3_7_markers_auroc) %>%
  ungroup() %>%
  arrange(sc3_7_markers_clusts, sc3_7_markers_auroc) %>%
  mutate(order = row_number())
  #arrange(sc3_7_markers_clusts, desc(sc3_7_markers_auroc)) %>%
#temp$feature_symbol <- factor(temp$feature_symbol, levels = temp$feature_symbol)
```

## Bar plot of highest 10 markers per cluster
```{r fig.height = 5, fig.width=6}
temp %>%
  ggplot(aes(x = order, y = sc3_7_markers_auroc, fill = factor(sc3_7_markers_clusts))) +
    geom_col(width = 0.8) +
    scale_y_continuous("AUROC", expand = c(0, 0), breaks = c(0.8), limits = c(0, 1)) +
    #labs(fill = "Adj p-value") +
    coord_flip() +
    facet_wrap(~sc3_7_markers_clusts, scales = "free", ncol = 4) +
    scale_fill_manual(values = RColorBrewer::brewer.pal(7, "Set2")) +
    scale_x_continuous(breaks = temp$order, labels = temp$feature_symbol, expand = c(0, 0)) +
    theme_bw(base_size = 16) +
    theme(panel.border = element_blank(), panel.grid = element_blank(), axis.line = element_line(colour = "black"), 
          axis.ticks.x = element_line(colour = "black"), axis.ticks.y = element_blank(), 
          axis.title.y = element_blank(), legend.position = "none", axis.text.x = element_text(size = 10, colour = "black"), 
          axis.text.y = element_text(colour = "black", size = 11), strip.background = element_blank(), 
          strip.text = element_text(size = 11, face = "italic"))
ggsave("plots/AUROC_plots/k7.pdf", dpi = 600, height = 5, width = 6)
```
