==> 10_Wikipaths_plots.Rmd <==
---
title: "10_wikipathways_ISC_plot"
author: "Jordi Camps"
date: "2018 M09 26"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyr)
library(dplyr)
library(ggplot2)
```

## Import text files
```{r}
go_isc1 = read.delim("tables/Wikipathways/WikiPathways_2016_table_ISC1.txt", header = TRUE)
go_isc2 = read.delim("tables/Wikipathways/WikiPathways_2016_table_ISC2.txt", header = TRUE)
go_isc3 = read.delim("tables/Wikipathways/WikiPathways_2016_table_ISC3.txt", header = TRUE)
go_fibro = read.delim("tables/Wikipathways/WikiPathways_2016_table_fibroblast.txt", header = TRUE)
go_smc = read.delim("tables/Wikipathways/WikiPathways_2016_table_SMC.txt", header = TRUE)
go_musc = read.delim("tables/Wikipathways/WikiPathways_2016_table_musc.txt", header = TRUE)
go_schwann = read.delim("tables/Wikipathways/WikiPathways_2016_table_schwann_cell.txt", header = TRUE)
head(go_isc1)
```

## Add column specifying type of cluster
```{r}
go_isc1$Cluster <- "ISC1"
go_isc2$Cluster <- "ISC2"
go_isc3$Cluster <- "ISC3"
go_fibro$Cluster <- "Fibroblast"
go_smc$Cluster <- "SMC"
go_musc$Cluster <- "musc"
go_schwann$Cluster <- "Schwann_cell"
```

## Bind dataframes together
```{r}
go <- rbind(go_isc1, go_isc2, go_isc3, go_fibro, go_smc, go_musc, go_schwann)
```

## Create term_clsut, split Term and filter for mus musculus
```{r}
go <- go %>%
  #mutate(Term_b = Term, Cluster_b = Cluster) %>%
  #tidyr::unite(Term_clust, Term_b, Cluster_b, sep = "_") %>%
  tidyr::separate(Term, c("Term", "Recipient", "Number"), sep = "_") %>%
  filter(Recipient == "Mus musculus")
```

## Order terms per group in descending combined score
```{r}
go$Cluster <- factor(go$Cluster, levels = c("ISC1", "ISC2", "ISC3", "Fibroblast", "SMC", "musc", "Schwann_cell"))
#go$Term_clust <- factor(go$Term_clust, levels = rev(go$Term_clust))
```

## Select only op 5 terms per  cluster and order by Combined score
```{r}
p <- tbl_df(go) %>%
  group_by(Cluster) %>%
  top_n(n = 5, wt = Combined.Score) %>%
  ungroup() %>%
  arrange(Cluster, desc(Combined.Score)) %>%
  mutate(order = row_number())
```

## Plot
```{r fig.height=7.5, fig.width=6.5}
p %>%
ggplot(aes(y = Combined.Score, x = rev(order), fill = Cluster)) +
  geom_col(width = 0.8) + 
  #geom_hline(yintercept = 1.3, color = "red", size = 1) +
  #scale_x_discrete(labels = rev(go$Term)) +
  scale_fill_manual(values = RColorBrewer::brewer.pal(7, "Set2")) +
  scale_x_continuous(breaks = p$order, labels = rev(p$Term), expand = c(0, 0)) +
  facet_wrap(~Cluster, ncol = 1, scales = "free_y") +
  coord_flip() +
  scale_y_continuous("Combined score", position = "left", expand = c(0, 0)) +
  theme_bw(base_size = 16) +
  theme(panel.border = element_blank(), panel.grid = element_blank(), axis.title.y = element_blank(), 
        axis.title.x = element_text(hjust = 1, size = 11),
        axis.ticks.y = element_blank(), axis.line = element_line(color = "black"), axis.text = element_text(color = "black", size = 11),
        axis.ticks.x = element_line(color = "black"), axis.text.y = element_text(hjust = 1, vjust = 0.3), strip.background = element_blank(),
        legend.position = "none", strip.text = element_blank())
ggsave("plots/GO/Wikipaths.pdf", dpi = 600, width = 6.5, height = 7.5)
```
