==> 12_stem_cell_population_markers_literature.Rmd <==
---
title: "progenitor markers"
author: "Jordi Camps"
date: "2018 M05 29"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(SingleCellExperiment)
library(scTools)
library(scater)
library(dplyr)
library(RColorBrewer)
library(tidyr)
set.seed(1234567)
```

```{r}
sce <- readRDS("data/reads_qc_scran_sc3.rds")
sce
```

Are factor already named?
```{r}
unique(sce$sc3_7_clusters)
```

rename populations
```{r}
sce$sc3_7_clusters <- plyr::revalue(sce$sc3_7_clusters, c("1" = "ISC 3", "2" = "ISC 1", "3" = "ISC 2", "4" = "Satellite cell", "5" = "Schwann cell", "6" = "SMC", "7" = "Fibroblast"))
sce$sc3_7_clusters <- factor(sce$sc3_7_clusters, levels = c("ISC 1", "ISC 2", "ISC 3", "Fibroblast", "SMC", "Satellite cell", "Schwann cell"))
```

## Check tSNE perplexity
```{r fig.height=3, fig.width=4}
for (p in c(5, 10, 15, 20, 25, 30, 35)) {
  print(plotTSNE(sce, colour_by = "sc3_7_clusters", return_SCE = FALSE, perplexity = p, exprs_values = "logcounts", rand_seed = 123456) + ggtitle(paste("Perplexity = ", p, sep = "")))
}
```

##Save tSNE into reduceddim
```{r fig.height=3, fig.width=4}
reducedDim(sce) <- NULL
sce <- plotTSNE(sce, colour_by = "genotype", return_SCE = TRUE, perplexity = 10, exprs_values = "logcounts", rand_seed = 123456)
sce$Dim1 <- reducedDims(sce)$TSNE[,1]
sce$Dim2 <- reducedDims(sce)$TSNE[,2]
```

Count overlapping markers
```{r}
genes <- c("Ly6a", "Pdgfra", "Alpl", "Peg3")
group = "sc3_7_clusters"

  rowData <- NULL
  for (gene in genes) {
    if (gene %in% row.names(sce)) {
      rowData[[gene]]  <- logcounts(sce)[gene, ]
    }
    else {
      print(paste0(gene, " not expressed or written wrong!", sep = ""))
    }
  }
    
rowData <- data.frame(rowData)
colData <- data.frame(cluster = sce[[group]])
temp <- cbind(rowData, colData)
#temp <- tidyr::gather(temp, gene, logcounts, -cluster)
#temp$gene <- factor(temp$gene, levels = genes)
```

```{r}
temp %>%
  tidyr::gather(gene, logcounts, -cluster) %>%
  ggplot(aes(x = logcounts)) +
  geom_histogram() +
  facet_wrap(~gene, scales = "free_y") +
  geom_vline(xintercept = 3.5, col = "red")
```

Ratio of positive cells per cluster per gene
```{r fig.height=3, fig.width=4.5}
temp <- temp %>% 
  tibble::rownames_to_column() %>%
  tidyr::gather(gene, logcounts, -cluster, -rowname)
```

```{r fig.height=3, fig.width=4.5}
temp$gene <- factor(temp$gene, levels = genes)
temp %>%
  group_by(cluster, gene) %>%
  summarise(n_events = n(), positive = sum(logcounts >= 3.5), p_events = positive / n_events) %>%
  ggplot(aes(x = cluster, y = p_events, fill = cluster)) +
  geom_col() +
  facet_wrap(~gene, scales = "free_x") +
  scale_y_continuous("Ratio positive cells", expand = c(0, 0), breaks = c(0.25, 0.5, 0.75, 1), labels = c("25%", "50%", "75%", "100%")) +
  scale_fill_manual(values = RColorBrewer::brewer.pal(7, "Set2")) +
  theme_bw(base_size = 16) +
  theme(panel.border = element_blank(), panel.grid = element_blank(), strip.background = element_blank(), axis.title.x = element_blank(),
        axis.line = element_line(colour = "black"), axis.ticks.y = element_line(colour = "black"), axis.ticks.x = element_blank(),
        axis.text.x = element_blank(), axis.text.y = element_text(colour = "black", size = 10), legend.title = element_blank())
ggsave("plots/Stem_cell_markers_comparison/Ratio_positive_cells_per_cluster.pdf", dpi = 600, height = 3, width = 4.5)
```

Ratio of positive cells per cluster
```{r fig.height=2.5, fig.width=4}
library(lazyeval)
df <- data_frame(a = c("Ly6a", "Ly6a", "Ly6a", "Pdgfra", "Pdgfra", "Alpl"),
                    b = c("Pdgfra", "Alpl", "Peg3", "Alpl", "Peg3", "Peg3"))

i <- 1

for (i in 1:6) {
print(temp %>%
  tidyr::spread(gene, logcounts) %>%
  group_by(cluster, rowname) %>%
  mutate_(.dots = setNames(list(lazyeval::interp(~ x >= 3.5 & y < 3.5, x = as.name(df[[i, 1]]), y = as.name(df[[i, 2]]))), "a")) %>%
  mutate_(.dots = setNames(list(lazyeval::interp(~ x < 3.5 & y >= 3.5, x = as.name(df[[i, 1]]), y = as.name(df[[i, 2]]))), "b")) %>%
  mutate_(.dots = setNames(list(lazyeval::interp(~ x >= 3.5 & y >= 3.5, x = as.name(df[[i, 1]]), y = as.name(df[[i, 2]]))), "ab")) %>%
  ungroup() %>%
  group_by(cluster) %>%
  summarise(a = sum(a) / n(), b = sum(b) / n(), ab = sum(ab) / n()) %>%
  tidyr::gather(marker, amount, -cluster) %>%
  ggplot(aes(x = cluster, y = amount, fill = marker)) +
  geom_col() +
  scale_y_continuous("Ratio positive cells", expand = c(0, 0), breaks = c(0.25, 0.5, 0.75, 1), labels = c("25%", "50%", "75%", "100%")) +
  scale_fill_manual(values = RColorBrewer::brewer.pal(3, "Set1"), labels = c(paste(df[i, 1], "+", df[i, 2], "-", sep = ""),
                                                                             paste(df[i, 1], "+", df[i, 2], "+", sep = ""),
                                                                             paste(df[i, 1], "-", df[i, 2], "+", sep = ""))) +
  theme_bw(base_size = 16) +
  theme(panel.border = element_blank(), panel.grid = element_blank(), strip.background = element_blank(), axis.title.x = element_blank(),
        axis.line = element_line(colour = "black"), axis.ticks.y = element_line(colour = "black"), axis.title.y = element_text(hjust = 1),
        axis.text.x = element_text(colour = "black", angle = 45, hjust = 1, size = 12), axis.text.y = element_text(colour = "black", size = 10), legend.title = element_blank()))
  ggsave(paste("plots/Stem_cell_markers_comparison/Ratio_", df[i, 1], "_", df[[i, 2]], ".pdf", sep = ""), dpi = 600, height = 2.5, width = 4)
}
```


Markers expression over clusters (boxplot)
```{r fig.height=2.5, fig.width=3}
genes <- c("Ly6a", "Pdgfra", "Alpl", "Peg3")

for (g in  genes) {
print(plot_expression(sce_object = sce, var = g, group = "sc3_7_clusters", facet = "horizontal", type = "boxplot") +
    scale_fill_manual(values = brewer.pal(7, "Set2")) +
      scale_y_continuous(expand = c(0, 0)) +
    theme(panel.border = element_blank(), strip.background = element_blank(), legend.position = "none", axis.line = element_line(colour = "black"),
          axis.ticks = element_line(colour = "black"), axis.text = element_text(colour = "black"), 
          axis.text.x = element_text(colour = "black", size = 10, angle = 45, hjust = 1), strip.text = element_text(colour = "black", size = 18))
    )
  ggsave(paste("plots/expression/", g, "_k7_boxplot.pdf", sep = ""), dpi = 600, height = 2.5, width = 3)
}
```

Markers expression  over clusters (violin plot)
```{r fig.height=2.5, fig.width=3}
genes <- c("Ly6a", "Pdgfra", "Alpl", "Peg3")

for (g in  genes) {
print(plot_expression(sce_object = sce, var = g, group = "sc3_7_clusters", facet = "horizontal", type = "violin") +
    stat_summary(fun.y = median, geom = "point", size = 2) +
    scale_fill_manual(values = brewer.pal(7, "Set2")) +
    scale_color_manual(values = brewer.pal(7, "Set2")) +
    scale_y_continuous(expand = c(0, 0)) +
    theme(panel.border = element_blank(), strip.background = element_blank(), legend.position = "none", axis.line = element_line(colour = "black"),
          axis.ticks = element_line(colour = "black"), axis.text = element_text(colour = "black"), 
          axis.text.x = element_text(colour = "black", size = 10, angle = 45, hjust = 1), strip.text = element_text(colour = "black", size = 18))
    )
  ggsave(paste("plots/expression/", g, "_k7_violin.pdf", sep = ""), dpi = 600, height = 2.5, width = 3)
}
```


## Statistics
### Wilcoxon test
```{r}
genes <- c("Ly6a", "Pdgfra", "Alpl", "Peg3")
group = "sc3_7_clusters"

  rowData <- NULL
  for (gene in genes) {
    if (gene %in% row.names(sce)) {
      rowData[[gene]]  <- logcounts(sce)[gene, ]
    }
    else {
      print(paste0(gene, " not expressed or written wrong!", sep = ""))
    }
  }
    
rowData <- data.frame(rowData)
colData <- data.frame(cluster = sce[[group]], cell = colnames(sce))
temp <- cbind(rowData, colData)
temp <- tidyr::gather(temp, gene, logcounts, -cluster, -cell)
temp$gene <- factor(temp$gene, levels = genes)
```

Ly6a 
```{r eval=FALSE, include=FALSE}
ind <- data_frame(a = c("ISC 1", "ISC 2", "ISC 3"), b = c("ISC 2", "ISC 3", "ISC 1"))
temp <- as_data_frame(temp)

for (g in genes) {
  for (i in 1:3) {
    p <- temp %>%
      filter(gene == g) %>%
      filter(cluster == ind[i, 1] | cluster == ind[i, 2])
print(paste("parameters: gene =", g, "clusters =", ind[i, 1], ind[i, 2], sep = " "))
print(wilcox.test(p$logcounts~p$cluster))
  }
}
```

### Correlation between markers
```{r fig.height=4, fig.width=4.5}
g1 <- c("Ly6a", "Ly6a", "Ly6a", "Pdgfra", "Pdgfra", "Pdgfra", "Peg3", "Peg3", "Peg3", "Alpl", "Alpl", "Alpl")
g2 <- c("Pdgfra", "Peg3", "Alpl", "Ly6a", "Peg3", "Alpl", "Ly6a", "Pdgfra", "Alpl", "Ly6a", "Pdgfra", "Peg3")
rho_data <- list()
temp_sub <- temp %>%
  spread(gene, logcounts)

for (i in 1:12) {
  rho <- cor.test(temp_sub[, g1[i]], temp_sub[, g2[i]], method = "spearman")
  rho_data[[i]] <- c(estimate = rho$estimate, p = rho$p.value, gene_1 = g1[i], gene_2 = g2[i])
}
df <- as.data.frame(t(as.data.frame(rho_data)))
df$estimate.rho <- as.numeric(df$estimate.rho)
df$p <- as.numeric(df$p)
head(df)

p <- df %>%
  dplyr::mutate(sign = p <= 0.05) %>%
  ungroup() %>%
  dplyr::arrange(gene_1, desc(estimate.rho)) %>%
  dplyr::mutate(order = row_number())
ggplot(p, aes(x = order, y = estimate.rho, fill = sign)) +
  geom_col(width = 0.8) +
  facet_wrap(~gene_1, scales = "free_x") +
  scale_fill_manual(values = c("grey80", "black"), labels = c("p > 0.05", "p < 0.05")) +
  scale_y_continuous("Spearman's rho", expand = c(0, 0)) +
  scale_x_continuous(breaks = p$order, labels = p$gene_2) +
  theme_bw(base_size = 16) +
  theme(panel.border = element_blank(), panel.grid = element_blank(), axis.line = element_line(colour = "black"), axis.ticks.x = element_blank(),
        axis.ticks.y = element_line(colour = "black"), axis.text = element_text(colour = "black"), axis.title.x = element_blank(), 
        legend.title = element_blank(), strip.background = element_blank(), strip.text = element_text(face = "italic"), 
        axis.text.x = element_text(angle = 45, hjust = 1))
ggsave("plots/Stem_cell_markers_comparison/pearson_corr.pdf", dpi = 600)
```
